rename social-icons to prevent adblockers from blocking scripts, fix for #491
