+++ webroot/styles/user-content.css
@@ -11,7 +11,7 @@
   --icon-width: 40px;
   height: var(--icon-width);
   width: var(--icon-width);
-  background-image: url(/img/social-icons.gif);
+  background-image: url(/img/platform-logos.gif);
   background-position: calc(var(--imgCol) * var(--icon-width)) calc(var(--imgRow) * var(--icon-width));
   transform: scale(.65);
 }
